Other section adjustments
