+++ src/views/OrganizationView.tsx
@@ -6,12 +6,13 @@ import { semibold } from '../fonts';
 import { Container } from 'react-bootstrap';
 import { Persona } from 'office-ui-fabric-react/lib/Persona';
 import { FontSizes } from '@fluentui/theme';
-import { Card, CardSection, ICardTokens } from "@uifabric/react-cards";
+import { Card, ICardTokens } from "@uifabric/react-cards";
 import { useTheme } from '@fluentui/react-theme-provider';
 import { getAdmins } from '../services/Requests';
 import { Separator } from '@fluentui/react/lib/Separator';
-import { getGroupsLength, getCdlsLength } from '../services/Requests';
 import { getCanMembers } from '../services/Requests';
+import { CompoundButton } from '@fluentui/react/lib/Button';
+import { IIconProps } from '@fluentui/react';
 
 const OrganizationView = () => {
     var theme = useTheme();
@@ -19,12 +20,10 @@ const OrganizationView = () => {
     const sectionCard = { minHeight: '130px', height: '100%', width: '100%', maxWidth: 'none', maxHeight: 'none' };
     const iconStyle = { color: theme.palette.themePrimary, fontSize: FontSizes.size24 };
     const iconStyle2 = { color: theme.palette.themePrimary };
-    const numberStyle = { color: theme.palette.themePrimary };
-    const mainCard = { minHeight: '160px', height: '100%', width: '100%', maxWidth: 'none', maxHeight: 'none', boxShadow: theme.effects.elevation8 };
     const admins = getAdmins();
     const canMembers = getCanMembers();
-    const groupsNumber = getGroupsLength();
-    const cdlsNumber = getCdlsLength();
+
+    const icon: IIconProps = { iconName: 'DocumentSearch' };
 
     return (
         <Container className="organization text-center">
@@ -36,55 +35,10 @@ const OrganizationView = () => {
                     </Text>
                 </div>
 
-                <div className="mb-3">
-                    <Row className="justify-content-center">
-                        <Col className="mb-3" xl={3} lg={4} md={4} sm={6} xs={12}>
-                            <Card tokens={cardTokens} style={mainCard} className="justify-content-center text-center">
-                                <CardSection>
-                                    <Icon iconName="UserOptional" style={iconStyle} />
-                                    <Text variant="large" className="mt-0">
-                                        Abbiamo più di<br />
-                                        <Text variant="xLarge" style={numberStyle}>2.000</Text> <br />
-                                        utenti
-                                    </Text>
-                                </CardSection>
-                            </Card>
-                        </Col>
-                        <Col className="mb-3" xl={3} lg={4} md={4} sm={6} xs={12}>
-                            <Card tokens={cardTokens} style={mainCard} className="justify-content-center text-center">
-                                <CardSection>
-                                    <Icon iconName="PeopleAlert" style={iconStyle} />
-                                    <Text variant="large" className="mt-0">
-                                        Abbiamo<br />
-                                        <Text variant="xLarge" style={numberStyle}>{groupsNumber}</Text> <br />
-                                        gruppi telegram
-                                    </Text>
-                                </CardSection>
-                            </Card>
-                        </Col>
-                        <Col className="mb-3" xl={3} lg={4} md={4} sm={6} xs={12}>
-                            <Card tokens={cardTokens} style={mainCard} className="justify-content-center text-center">
-                                <CardSection>
-                                    <Icon iconName="PublishCourse" style={iconStyle} />
-                                    <Text variant="large" className="mt-0">
-                                        Copriamo<br />
-                                        <Text variant="xLarge" style={numberStyle}>{cdlsNumber}</Text> <br />
-                                        corsi di laurea
-                                    </Text>
-                                </CardSection>
-                            </Card>
-                        </Col>
-                    </Row>
-                </div>
-
                 <div className="mb-2">
-                    <Text variant="medium">
-                        <div className="mb-0">
-                            <Icon iconName="ComplianceAudit" className="mr-1" style={{ color: theme.palette.themePrimary, fontSize: '14px' }} />
-                            Puoi trovare il nostro statuto cliccando <Link href="https://github.com/StudentiUniMi/docs/blob/main/statuto.md" target="_blank">qui</Link>.
-                        </div>
-                        Di seguito è possibile trovare i membri che hanno un ruolo attivo nel network e i relativi contatti.
-                    </Text>
+                    <CompoundButton primary secondaryText="Dai un'occhiata al nostro statuto!" href="https://github.com/StudentiUniMi/docs/blob/main/statuto.md" target="_blank" style={{ textDecoration: 'none' }} iconProps={icon}>
+                        Statuto
+                    </CompoundButton>
                 </div>
             </div>
 
@@ -143,7 +97,7 @@ const OrganizationView = () => {
 
 
             <div className="mb-4">
-                <div className="mb-3"><Separator><Text variant="large" styles={semibold}>Amministratori e moderatori</Text></Separator></div>
+                <div className="mb-3"><Separator><Text variant="large" styles={semibold}>Amministratori dei gruppi telegram</Text></Separator></div>
 
                 <Row className="justify-content-center">
                     {
